rewrite: WIP rewrite to version 2. Fix bugs found during testing of the article endpoint.
