+++ tests/api/test_article.py
@@ -14,7 +14,7 @@ logger = logging.getLogger(__name__)
 # TestCase.maxDiff = None
 
 
-class TestGetArticleStatistics(TestCase):
+class TestArticle(TestCase):
     """Generated by ChatGPT using this prompt:
         Create a test for this flask api endpoint
     import logging
