+++ tests/deprecated/test_get_url_statistics.py
