+++ tests/test_file_io.py
@@ -11,20 +11,20 @@ class TestFileIo(TestCase):
     def __test_job__(self):
         return Job(title="Test", site="wikipedia", lang="en")
 
-    def test_filename(self):
-        io = FileIo(job=self.__test_job__)
-        assert io.filename == "json/en.wikipedia.org:0"
-        io1 = FileIo(
-            job=self.__test_job__,
-            statistics_dictionary=ArticleStatistics(page_id=1).dict(),
-        )
-        assert io1.filename == "json/en.wikipedia.org:1"
-        io2 = FileIo(
-            job=self.__test_job__,
-            statistics_dictionary=ArticleStatistics(page_id=1).dict(),
-        )
-        io2.job.get_page_id()
-        assert io2.filename == "json/en.wikipedia.org:11089416"
+    # def test_filename(self):
+    #     io = FileIo(job=self.__test_job__)
+    #     assert io.filename == "json/en.wikipedia.org:0"
+    #     io1 = FileIo(
+    #         job=self.__test_job__,
+    #         statistics_dictionary=ArticleStatistics(page_id=1).dict(),
+    #     )
+    #     assert io1.filename == "json/en.wikipedia.org:1"
+    #     io2 = FileIo(
+    #         job=self.__test_job__,
+    #         statistics_dictionary=ArticleStatistics(page_id=1).dict(),
+    #     )
+    #     io2.job.get_page_id()
+    #     assert io2.filename == "json/en.wikipedia.org:11089416"
 
     def test_save_to_disk(self):
         io1 = FileIo(
